+++ server/output/chapter02.h
@@ -1,0 +1,8 @@
+#ifndef CHAPTER02_H
+#define CHAPTER02_H
+
+#include <loke/scene.h>
+
+// No scenes yet
+
+#endif // CHAPTER02_H