+++ server/output/scene_forrest.c
@@ -4,5 +4,5 @@
 void scene_forrest(GameState* state) {
     SceneContext* ctx = get_current_context();
     scene_set_text(ctx, "Skoven ");
-    scene_add_option(ctx, "Continue", NULL, true);
+    scene_add_option(ctx, "Tag hjem", scene_home, true);
 }