+++ server/output/scene_hills.c
@@ -4,5 +4,5 @@
 void scene_hills(GameState* state) {
     SceneContext* ctx = get_current_context();
     scene_set_text(ctx, "Nogle bakker");
-    scene_add_option(ctx, "Continue", NULL, true);
+    scene_add_option(ctx, "Løb hjem", scene_home, true);
 }